refactor cardtype to fix frontend
